+++ packages/lib/env.js
@@ -1,8 +1,8 @@
 /**
  * Environment configuration object
- * @type {{mobxVersion: 4 | 5 | 6, compiler: string}}
+ * @type {{mobxVersion: 4 | 5 | 6 | 7, compiler: string}}
  */
 module.exports.env = {
-  mobxVersion: Number(process.env.MOBX_VERSION || "6"),
+  mobxVersion: Number(process.env.MOBX_VERSION || "7"),
   compiler: process.env.COMPILER || "tsc", // tsc | tsc-experimental-decorators | babel | swc
 }